Prototype fix

+++ include/input.h
@@ -16,7 +16,7 @@ extern "C" {
 
 int snd_input_stdio_open(snd_input_t **inputp, const char *file, const char *mode);
 int snd_input_stdio_attach(snd_input_t **inputp, FILE *fp, int _close);
-int snd_input_buffer_open(snd_input_t **inputp, const char *buffer, int size);
+int snd_input_buffer_open(snd_input_t **inputp, const char *buffer, ssize_t size);
 int snd_input_close(snd_input_t *input);
 int snd_input_scanf(snd_input_t *input, const char *format, ...) __attribute__ ((format (scanf, 2, 3)));
 char *snd_input_gets(snd_input_t *input, char *str, size_t size);
